business/requirements.txt
Django==3.1.1
djangorestframework
djangorestframework-simplejwt
pyjwt
psycopg2-binary
sqlparse